Ported Fortran code into CSharp
Ported the Fortran code into CSharp almost 1:1 with some differences in structure to account for different language features. Added a simulation object and added the ability to write results into a file when enabled. I compared the results of Fortran and CSharp and after almost 30,000 simulation steps the differences are still negligible (within rounding error).